[tests] Consolidate production CLI coverage
Keep the production-binary protocol regression in the behavior-oriented
CLI harness instead of building a one-test integration artifact.

gherrit-pr-id: Gh37h43czhwq7j5f2u5ioe2j2dnxyp36a
