+++ tests/cli/main.rs
@@ -2,3 +2,4 @@ mod commit_msg;
 mod install;
 mod manage;
 mod post_checkout;
+mod production;
